+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add support for Finnish language
 - Add script for packing project
 - Add scripts for managing translations
+- Add age check before registering
 
 ### Changed
 - Hide all other links except account and password